trivia($locale): update default locale
